# Check use of invalidated resources at runtime