Work towards https://github.com/dapperlabs/cadence-private-issues/issues/19

## Description

This adds second layer of runtime validation to check use of invalidated (moved/destroyed) resources. This check relies on:
- An external mapping, which keeps the resource-to-variable mapping.
- Assumes there could be at-most one mapping for each resource value at any given time.

______

<!-- Complete: -->

- [x] Targeted PR against `master` branch
- [x] Linked to Github issue with discussion and accepted design OR link to spec that describes this work
- [x] Code follows the [standards mentioned here](https://github.com/onflow/cadence/blob/master/CONTRIBUTING.md#styleguides)
- [x] Updated relevant documentation 
- [x] Re-reviewed `Files changed` in the Github PR explorer
- [x] Added appropriate labels