@turbolent 2022-03-01
@SupunS Could you please check performance locally? I'd be curious to see if it's better than what CI says the current benchstat numbers above are not quite reliable, right? (e.g. parsing is supposedly slower) Allocation increase is not that big a deal, but would be good to keep overhead <~10%