@SupunS 2022-03-01
Yeah the CI results are kind of not accurate. here. e.g: `QualifiedIdentifierCreation` is just testing a utility method that doesn't run any cadence code, and has no impact from the change.

local benchmark results:
```
name                                              old time/op    new time/op    delta
pkg:github.com/onflow/cadence/runtime goos:darwin goarch:amd64
RuntimeFungibleTokenTransfer-8                       599µs ± 4%     599µs ± 5%   ~     (p=1.000 n=3+3)
RuntimeResourceDictionaryValues-8                   9.02ms ± 1%    8.77ms ± 1%   ~     (p=0.100 n=3+3)
pkg:github.com/onflow/cadence/runtime/parser2 goos:darwin goarch:amd64
ParseDeploy/byte_array-8                            13.6ms ± 0%    14.4ms ±12%   ~     (p=1.000 n=3+3)
ParseDeploy/decode_hex-8                             841µs ± 0%     864µs ± 5%   ~     (p=0.200 n=3+3)
ParseFungibleToken-8                                 143µs ± 1%     144µs ± 3%   ~     (p=0.400 n=3+3)
ParseInfix-8                                        6.58µs ± 2%    6.53µs ± 1%   ~     (p=1.000 n=3+3)
ParseArray-8                                        9.18ms ± 4%    8.90ms ± 1%   ~     (p=0.700 n=3+3)
pkg:github.com/onflow/cadence/runtime/sema goos:darwin goarch:amd64
QualifiedIdentifierCreation/One_level-8             1.95ns ± 2%    1.94ns ± 2%   ~     (p=0.400 n=3+3)
QualifiedIdentifierCreation/Three_levels-8          88.1ns ± 1%    87.2ns ± 1%   ~     (p=0.400 n=3+3)
pkg:github.com/onflow/cadence/runtime/tests/checker goos:darwin goarch:amd64
ContractInterfaceFungibleToken-8                    31.2µs ± 2%    30.8µs ± 2%   ~     (p=0.400 n=3+3)
CheckContractInterfaceFungibleTokenConformance-8     101µs ± 1%     100µs ± 1%   ~     (p=0.400 n=3+3)
pkg:github.com/onflow/cadence/runtime/tests/interpreter goos:darwin goarch:amd64
InterpretRecursionFib-8                             1.95ms ± 2%    2.05ms ± 2%   ~     (p=0.100 n=3+3)
NewInterpreter/new_interpreter-8                     833ns ± 1%     883ns ± 3%   ~     (p=0.100 n=3+3)
NewInterpreter/new_sub-interpreter-8                1.67µs ± 2%    1.72µs ± 1%   ~     (p=0.100 n=3+3)

name                                              old alloc/op   new alloc/op   delta
pkg:github.com/onflow/cadence/runtime goos:darwin goarch:amd64
RuntimeFungibleTokenTransfer-8                       273kB ± 0%     273kB ± 0%   ~     (p=0.100 n=3+3)
RuntimeResourceDictionaryValues-8                   4.05MB ± 0%    4.05MB ± 0%   ~     (p=1.000 n=3+3)
pkg:github.com/onflow/cadence/runtime/sema goos:darwin goarch:amd64
QualifiedIdentifierCreation/One_level-8              0.00B          0.00B        ~     (all equal)
QualifiedIdentifierCreation/Three_levels-8           64.0B ± 0%     64.0B ± 0%   ~     (all equal)
pkg:github.com/onflow/cadence/runtime/tests/checker goos:darwin goarch:amd64
ContractInterfaceFungibleToken-8                    26.6kB ± 0%    26.6kB ± 0%   ~     (all equal)
CheckContractInterfaceFungibleTokenConformance-8    66.2kB ± 0%    66.2kB ± 0%   ~     (p=1.000 n=3+3)
pkg:github.com/onflow/cadence/runtime/tests/interpreter goos:darwin goarch:amd64
InterpretRecursionFib-8                             1.24MB ± 0%    1.24MB ± 0%   ~     (p=0.800 n=3+3)
NewInterpreter/new_interpreter-8                      768B ± 0%      848B ± 0%   ~     (p=0.100 n=3+3)
NewInterpreter/new_sub-interpreter-8                1.24kB ± 0%    1.32kB ± 0%   ~     (p=0.100 n=3+3)

name                                              old allocs/op  new allocs/op  delta
pkg:github.com/onflow/cadence/runtime goos:darwin goarch:amd64
RuntimeFungibleTokenTransfer-8                       4.53k ± 0%     4.53k ± 0%   ~     (p=0.100 n=3+3)
RuntimeResourceDictionaryValues-8                     102k ± 0%      102k ± 0%   ~     (p=0.400 n=3+3)
pkg:github.com/onflow/cadence/runtime/sema goos:darwin goarch:amd64
QualifiedIdentifierCreation/One_level-8               0.00           0.00        ~     (all equal)
QualifiedIdentifierCreation/Three_levels-8            2.00 ± 0%      2.00 ± 0%   ~     (all equal)
pkg:github.com/onflow/cadence/runtime/tests/checker goos:darwin goarch:amd64
ContractInterfaceFungibleToken-8                       458 ± 0%       458 ± 0%   ~     (all equal)
CheckContractInterfaceFungibleTokenConformance-8     1.07k ± 0%     1.07k ± 0%   ~     (all equal)
pkg:github.com/onflow/cadence/runtime/tests/interpreter goos:darwin goarch:amd64
InterpretRecursionFib-8                              25.0k ± 0%     25.0k ± 0%   ~     (all equal)
NewInterpreter/new_interpreter-8                      12.0 ± 0%      13.0 ± 0%   ~     (p=0.100 n=3+3)
NewInterpreter/new_sub-interpreter-8                  38.0 ± 0%      39.0 ± 0%   ~     (p=0.100 n=3+3)
```